Clean up JSON classes
