+++ JsonClasses/Anime.cs
@@ -2,5 +2,11 @@
 
 namespace SeiyuuSync.JsonClasses
 {
-    
+    public class Anime
+    {
+        [JsonPropertyName("id")]
+        public int Id { get; set; }
+        [JsonPropertyName("title")]
+        public string Title { get; set; }
+    }
 }
