+++ JsonClasses/Node.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Text.Json.Serialization;
+using System.Threading.Tasks;
+
+namespace SeiyuuSync.JsonClasses
+{
+    public class Node
+    {
+        [JsonPropertyName("node")]
+        public Anime Anime { get; set; }
+    }
+}
